Added some rough Hadoop scripts, removed some cruft
