--- tasks/access/templates/site-target-template.md
@@ -1,3 +1,0 @@
-{{ json }}
-
-{{ description }}
