--- tasks/access/templates/title-level-template.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<OAI-PMH xmlns="http://www.openarchives.org/OAI/2.0/">
-  <ListRecords>
-      {% for record in records %}
-      <record>
-          <header>
-              <identifier>{{ record.id }}</identifier>
-          </header>
-          <metadata>
-              <oai_dc:dc xmlns:oai_dc="http://www.openarchives.org/OAI/2.0/oai_dc/">
-                  <dc:source xmlns:dc="http://purl.org/dc/elements/1.1/">{{ record.url }}</dc:source>
-                  <dc:publisher xmlns:dc="http://purl.org/dc/elements/1.1/">{{ record.publisher }}</dc:publisher>
-                  <dc:title xmlns:dc="http://purl.org/dc/elements/1.1/">{{ record.title }}</dc:title>
-                  <dc:date xmlns:dc="http://purl.org/dc/elements/1.1/">{{ record.date }}</dc:date>
-                  <dc:rights xmlns:dc="http://purl.org/dc/elements/1.1/">{{ record.rights }}</dc:rights>
-                  {% if record.subject %}
-                  <dc:subject xmlns:dc="http://purl.org/dc/elements/1.1/">{{ record.subject }}</dc:subject>
-                  {% endif %}
-                  <xlink:href xmlns:xlink="http://www.w3.org/1999/xlink">{{ record.wayback_url }}</xlink:href>  
-              </oai_dc:dc>
-          </metadata>
-      </record>
-      {% endfor %}
-  </ListRecords>
-</OAI-PMH>
